try to get syscall prototype declarations working

+++ sysdeps/unix/sysv/zos/s390/s390-64/linux_shims/open.c
@@ -4,6 +4,7 @@
 #include <string.h>
 #include <limits.h>
 #include <sys/cdefs.h>
+#include <sys/types.h>
 #include "linux_syscall_shim.h"
 
 typedef void (*bpx4opn_t)(const uint32_t *pathname_len,
@@ -22,8 +23,8 @@ typedef void (*bpx4opn_t)(const uint32_t *pathname_len,
    standard wrong, but use it anyway for now.
  */
 
-int32_t
-__linux_compat_open (long *err_code, const char *pathname, int flags, mode_t mode)
+int
+__linux_compat_open (int *err_code, const char *pathname, int flags, mode_t mode)
 {
   int32_t retval, reason_code;
   int32_t pathname_len = SAFE_PATHLEN_OR_FAIL_WITH(pathname, -1);
@@ -33,8 +34,8 @@ __linux_compat_open (long *err_code, const char *pathname, int flags, mode_t mod
   return retval;
 }
 
-int32_t
-__linux_compat_openat (long *err_code, int dirfd, const char *pathname, int flags,
+int
+__linux_compat_openat (int *err_code, int dirfd, const char *pathname, int flags,
 		       mode_t mode)
 {
   if (dirfd == AT_FDCWD || *pathname == '/')
